requirements.txt
# Core dependencies
pydantic>=2.0.0
pathlib
python-dotenv>=1.0.0

# LangGraph ecosystem (latest versions)
langgraph>=0.5.2
langgraph-prebuilt>=0.5.2
langchain>=0.3.0
langchain-ollama>=0.3.0

# LangGraph checkpointers for state persistence
langgraph-checkpoint>=1.0.0
langgraph-checkpoint-sqlite>=1.0.0  # For development/testing
langgraph-checkpoint-postgres>=1.0.0  # For production (optional)

# Ollama client (latest available version)
ollama>=0.5.1

# System monitoring
psutil>=5.9.0

# Optional: For future enhancements
# sentence-transformers>=2.2.0  # For better embeddings
# faiss-cpu>=1.7.0              # For vector similarity search
# numpy>=1.21.0                 # For numerical operations
